Style: Fix the code in accordance to cpplint
Fix/Ignored 70 style issues according to cpplint, majorly styles and
formatting issues. This commit also adds const keyword to references
inputs of the following method and function:
KalmanFilter::Init
tools.cpp/check_files

cpplint is a tool for linting C++ code according to Google C++ style
guide, and can be installed using:
```
pip install cpplint
```

+++ src/measurement_package.h
@@ -4,10 +4,10 @@
 #include "Eigen/Dense"
 
 class MeasurementPackage {
-public:
-  long long timestamp_;
+ public:
+  uint64_t timestamp_;
 
-  enum SensorType{
+  enum SensorType {
     LASER,
     RADAR
   } sensor_type_;
